Give now inserts into first available slot, however the items do not stack, as we need a new method for max stacking.
All methods concerning inventory should be moved to Inventory.cs. No calls should be made to Player.cs for inventory except to get the Inventory handle. OR if you know what your doing.

We also need to workout a way to handle all the little changes the server needs to watch to have an accurate inventory. which we need to show what item the player is holding. we also need exceptions to the item id's, since blocks can be held as items.

Ill be gone for a few days, i expect something to get done :3 (not in this list necessarily, but overall, i know bizzare is working on the terrain gen :D)