Added a couple new things.
Added PlaceButton blockchange, confirmed working, it only needs to check the block that its being placed in to be sure its air.

Added a few new methods for working with blockchanges, and a new enum *Directions* in DataValues

in datavalues.cs there is now a class that can be used to store block information and methods for working with blocks. the method thats in there now is just used for checking if a block is a true block or not.